packages/kalshi-plugin/pyproject.toml
[project]
name = "kalshi-plugin"
version = "0.1.0"
description = "OpenClaw Kalshi plugin scaffold"
requires-python = ">=3.9"
authors = [{ name = "Michael" }]
dependencies = []

[project.optional-dependencies]
test = [
  "pytest>=8.0"
]
dev = [
  "pytest>=8.0",
  "PyYAML>=6.0",
  "httpx>=0.27"
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
